refactor: relocate data utility module
